Moved Cell components into subfolder
