+++ src/components/List/Cells/ApplicantCell.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { CellProps } from '../../utils/types'
+import { CellProps } from '../../../utils/types'
 
 const ApplicantCell: React.FC<CellProps> = ({ application }) => {
   // TODO
