+++ src/components/List/Cells/ApplicationNameCell.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
-import { CellProps } from '../../utils/types'
+import { CellProps } from '../../../utils/types'
 
 const ApplicationNameCell: React.FC<CellProps> = ({ application }) => (
   <p>
