+++ src/components/List/Cells/ConsolidatorCell.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { CellProps } from '../../utils/types'
+import { CellProps } from '../../../utils/types'
 
 const ConsolidatorCell: React.FC<CellProps> = ({ application }) => {
   // TODO
