+++ src/components/List/Cells/LastActiveDateCell.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { DateTime } from 'luxon'
-import { CellProps } from '../../utils/types'
+import { CellProps } from '../../../utils/types'
 
 const LastActiveDateCell: React.FC<CellProps> = ({ application }) => {
   return <p>{DateTime.fromISO(application.lastActiveDate).toISODate()}</p>
