+++ src/components/List/Cells/OutcomeCell.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Segment } from 'semantic-ui-react'
-import { CellProps } from '../../utils/types'
+import { CellProps } from '../../../utils/types'
 
 const OutcomeCell: React.FC<CellProps> = ({ application }) => (
   <Segment basic textAlign="center">
