+++ src/components/List/Cells/SerialNumberCell.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { CellProps } from '../../utils/types'
+import { CellProps } from '../../../utils/types'
 
 const SerialNumberCell: React.FC<CellProps> = ({ application }) => {
   return <p>{application.serial}</p>
